embedding/requirements.txt
google-generativeai==0.3.2
voyageai==0.2.0
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0